:recycle: Refactor: Additional small code cleanups
