+++ src/components/Gallery/styles.ts
@@ -55,7 +55,7 @@ export const Modal = styled.div`
   width: 100%;
   height: 100%;
 
-  &.visivel {
+  &.is-visible {
     display: flex;
   }
