+++ src/utils/index.ts
@@ -1,0 +1,6 @@
+export const formatPrice = (preco = 0) => {
+  return new Intl.NumberFormat('pt-br', {
+    style: 'currency',
+    currency: 'BRL'
+  }).format(preco)
+}
